.devdev/aws-bedrock-nova-sonic-try/pyproject.toml
[project]
name = "aws-bedrock-nova-sonic-try"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore>=1.1.2",
    "strands-agents-tools>=0.2.19",
    "strands-agents[bidi,bidi-all]>=1.21.0",
]

[dependency-groups]
dev = [
    "bedrock-agentcore-starter-toolkit>=0.2.5",
]
